requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP Client
httpx>=0.25.0

# Environment Variables
python-dotenv>=1.0.0

# Bitcoin Utilities
bitcoin-utils>=0.7.0
